:recycle: separate parser logic with test code
